Feat: Rank라는 구조체 만들고 Combinaiton 이동
